feat: Consolidate demo activities
