+++ demo/README.md
@@ -1,0 +1,17 @@
+# PayPal Messages Demo
+
+This demo showcases the different ways to integrate the PayPal Messages component into an Android application.
+
+## Structure
+
+The demo is structured as follows:
+
+- `DemoActivity`: The main entry point of the demo application. It uses a `TabLayout` and `ViewPager2` to display the different integration methods.
+- `SectionsPagerAdapter`: An adapter that provides the fragments for the `ViewPager2`.
+- `XmlFragment`: A fragment that demonstrates how to use the `PayPalMessageView` in an XML layout.
+- `JetpackFragment`: A fragment that demonstrates how to use the `PayPalMessageView` in a Jetpack Compose UI.
+- `JetpackComposableFragment`: A fragment that demonstrates how to use the `PayPalComposableMessage` in a Jetpack Compose UI.
+
+## How to Run
+
+To run the demo, simply build and run the `demo` module on an Android device or emulator.
