+++ demo/src/main/AndroidManifest.xml
@@ -3,6 +3,7 @@
 	xmlns:tools="http://schemas.android.com/tools">
 
 	<application
+		android:name=".PayPalMessagesApplication"
 		android:allowBackup="true"
 		android:dataExtractionRules="@xml/data_extraction_rules"
 		android:fullBackupContent="@xml/backup_rules"
@@ -14,17 +15,7 @@
 		tools:targetApi="31">
 
 		<activity
-			android:name=".JetpackActivity"
-			android:exported="true"
-			android:theme="@style/Theme.AppCompat.Light.NoActionBar">
-			<intent-filter>
-				<action android:name="android.intent.action.MAIN" />
-				<category android:name="android.intent.category.LAUNCHER" />
-			</intent-filter>
-		</activity>
-
-		<activity
-			android:name=".XmlActivity"
+			android:name=".DemoActivity"
 			android:exported="true"
 			android:theme="@style/Theme.AppCompat.Light">
 			<intent-filter>
@@ -34,14 +25,8 @@
 		</activity>
 
 		<activity
-			android:name=".JetpackComposableActivity"
-			android:exported="true"
-			android:theme="@style/Theme.AppCompat.Light.NoActionBar">
-			<intent-filter>
-				<action android:name="android.intent.action.MAIN" />
-				<category android:name="android.intent.category.LAUNCHER" />
-			</intent-filter>
-		</activity>
-
+			android:name=".proguard.ProGuardTestActivity"
+			android:exported="false"
+			android:theme="@style/Theme.AppCompat.Light" />
 	</application>
 </manifest>
