+++ demo/src/main/java/com/paypal/messagesdemo/PageViewModel.kt
@@ -1,0 +1,18 @@
+package com.paypal.messagesdemo.ui.main
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.map
+
+class PageViewModel : ViewModel() {
+
+	private val _index = MutableLiveData<Int>()
+	val text: LiveData<String> = _index.map {
+		"Hello world from section: $it"
+	}
+
+	fun setIndex(index: Int) {
+		_index.value = index
+	}
+}
