+++ demo/src/main/java/com/paypal/messagesdemo/SectionsPagerAdapter.kt
@@ -1,0 +1,41 @@
+package com.paypal.messagesdemo.ui.main
+
+import android.content.Context
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentManager
+import androidx.fragment.app.FragmentPagerAdapter
+import com.paypal.messagesdemo.JetpackComposableFragment
+import com.paypal.messagesdemo.JetpackFragment
+import com.paypal.messagesdemo.R
+import com.paypal.messagesdemo.XmlFragment
+
+private val TAB_TITLES = arrayOf(
+	R.string.tab_text_1,
+	R.string.tab_text_2,
+	R.string.tab_text_3
+)
+
+/**
+ * A [FragmentPagerAdapter] that returns a fragment corresponding to
+ * one of the sections/tabs/pages.
+ */
+class SectionsPagerAdapter(private val context: Context, fm: FragmentManager) :
+	FragmentPagerAdapter(fm) {
+
+	override fun getItem(position: Int): Fragment {
+		return when (position) {
+			0 -> XmlFragment()
+			1 -> JetpackFragment()
+			2 -> JetpackComposableFragment()
+			else -> Fragment()
+		}
+	}
+
+	override fun getPageTitle(position: Int): CharSequence? {
+		return context.resources.getString(TAB_TITLES[position])
+	}
+
+	override fun getCount(): Int {
+		return 3
+	}
+}
